requirements.txt
opencv-python==4.8.1.78
Pillow==10.0.1
pytesseract==0.3.10
pdf2image==1.16.3
google-generativeai==0.3.2
python-dotenv==1.0.0
pydantic==2.5.0
docscan==1.0.5
numpy==1.24.3
scipy==1.11.4
imutils==0.5.4
